Cleaned up files, Updated S3 to scan in a lower cost manor, as well as the deep analysis mode still
